Enable OSGI tests
