+++ osgi-tests/src/test/java/io/siddhi/distribution/test/osgi/FileSystemPersistenceStoreTestcase.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.wso2.carbon.analytics.test.osgi;
+package io.siddhi.distribution.test.osgi;
 
+import io.siddhi.distribution.core.core.internal.StreamProcessorDataHolder;
 import org.apache.log4j.Logger;
 import org.awaitility.Awaitility;
 import org.ops4j.pax.exam.Configuration;
@@ -28,15 +29,16 @@ import org.osgi.framework.BundleContext;
 import org.testng.Assert;
 import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
-import org.wso2.carbon.analytics.test.osgi.util.SiddhiAppUtil;
+import io.siddhi.distribution.test.osgi.util.SiddhiAppUtil;
 import org.wso2.carbon.container.CarbonContainerFactory;
-import org.wso2.carbon.stream.processor.core.internal.StreamProcessorDataHolder;
 import org.wso2.siddhi.core.SiddhiAppRuntime;
 import org.wso2.siddhi.core.SiddhiManager;
+import org.wso2.siddhi.core.exception.CannotRestoreSiddhiAppStateException;
 
 import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.concurrent.TimeUnit;
 import javax.inject.Inject;
 
@@ -46,9 +48,9 @@ import static org.wso2.carbon.container.options.CarbonDistributionOption.copyFil
 @Listeners(PaxExam.class)
 @ExamReactorStrategy(PerClass.class)
 @ExamFactory(CarbonContainerFactory.class)
-public class FileSystemPersistenceStoreConfigTestcase {
+public class FileSystemPersistenceStoreTestcase {
 
-    private static final Logger log = Logger.getLogger(FileSystemPersistenceStoreConfigTestcase.class);
+    private static final Logger log = Logger.getLogger(FileSystemPersistenceStoreTestcase.class);
     private static final String DEPLOYMENT_FILENAME = "deployment.yaml";
     private static final String PERSISTENCE_FOLDER = "siddhi-app-persistence";
     private static final String SIDDHIAPP_NAME = "SiddhiAppPersistence";
@@ -65,8 +67,8 @@ public class FileSystemPersistenceStoreConfigTestcase {
         if (basedir == null) {
             basedir = Paths.get(".").toString();
         }
-        carbonYmlFilePath = Paths.get(basedir, "src", "test", "resources", "conf", "persistence", "file",
-                "configTest", DEPLOYMENT_FILENAME);
+        carbonYmlFilePath = Paths.get(basedir, "src", "test", "resources",
+                "conf", "persistence", "file", "default", DEPLOYMENT_FILENAME);
         return copyFile(carbonYmlFilePath, Paths.get("conf", "worker", DEPLOYMENT_FILENAME));
     }
 
@@ -76,13 +78,14 @@ public class FileSystemPersistenceStoreConfigTestcase {
         log.info("Running - "+ this.getClass().getName());
         return new Option[]{
                 copyCarbonYAMLOption(),
-                carbonDistribution(Paths.get("target", "wso2das-" +
-                        System.getProperty("carbon.analytic.version")), "worker")
+                carbonDistribution(
+                        Paths.get("target", "wso2sp-"
+                                + System.getProperty("io.siddhi.distribution.version")), "worker")
         };
     }
 
     @Test
-    public void testFileSystemPersistenceWithDefaultValues() throws InterruptedException {
+    public void testFileSystemPersistence() throws InterruptedException {
         Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> {
             SiddhiManager siddhiManager = StreamProcessorDataHolder.getSiddhiManager();
             return siddhiManager != null;
@@ -97,25 +100,72 @@ public class FileSystemPersistenceStoreConfigTestcase {
 
         final File[] file = {new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME)};
         Assert.assertEquals(file[0].exists() && file[0].isDirectory(), false, "No Folder Created");
-
         log.info("Waiting for first time interval for state persistence");
         Awaitility.await().atMost(2, TimeUnit.MINUTES).until(() -> {
             file[0] = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
             return file[0].exists() && file[0].isDirectory() && file[0].list().length == 1;
         });
-        log.info("Waiting for second time interval for state persistence");
-        Awaitility.await().atMost(2, TimeUnit.MINUTES).until(() -> {
-            file[0] = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
-            return file[0].exists() && file[0].isDirectory() && file[0].list().length == 2;
-        });
-        log.info("Waiting for third time interval for state persistence");
-        Awaitility.await().atMost(2, TimeUnit.MINUTES).until(() -> {
-            file[0] = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
-            return file[0].exists() && file[0].isDirectory() && file[0].list().length == 3;
-        });
 
         Assert.assertEquals(file[0].exists() && file[0].isDirectory(), true);
-        Assert.assertEquals(file[0].list().length, 3, "There should be three revisions persisted");
+        Assert.assertEquals(file[0].list().length, 1, "There should be one revision persisted");
 
     }
+
+    @Test(dependsOnMethods = {"testFileSystemPersistence"})
+    public void testRestoreFromFileSystem() throws InterruptedException {
+        log.info("Waiting for second time interval for state persistence");
+        Awaitility.await().atMost(2, TimeUnit.MINUTES).until(() -> {
+            File file = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
+            return file.exists() && file.isDirectory() && file.list().length == 2;
+        });
+
+        SiddhiManager siddhiManager = StreamProcessorDataHolder.getSiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.getSiddhiAppRuntime(SIDDHIAPP_NAME);
+        log.info("Shutting Down SiddhiApp");
+        siddhiAppRuntime.shutdown();
+        log.info("Creating New SiddhiApp with Same Name");
+        SiddhiAppRuntime newSiddhiAppRuntime = SiddhiAppUtil.
+                createSiddhiApp(StreamProcessorDataHolder.getSiddhiManager());
+        try {
+            newSiddhiAppRuntime.restoreLastRevision();
+        } catch (CannotRestoreSiddhiAppStateException e) {
+            Assert.fail("Restoring of Siddhi App Failed");
+        }
+
+        SiddhiAppUtil.sendDataToStream("WSO2", 280L, newSiddhiAppRuntime);
+        SiddhiAppUtil.sendDataToStream("WSO2", 150L, newSiddhiAppRuntime);
+        SiddhiAppUtil.sendDataToStream("WSO2", 200L, newSiddhiAppRuntime);
+        SiddhiAppUtil.sendDataToStream("WSO2", 270L, newSiddhiAppRuntime);
+        SiddhiAppUtil.sendDataToStream("WSO2", 280L, newSiddhiAppRuntime);
+
+        Assert.assertEquals(SiddhiAppUtil.outputElementsArray, Arrays.asList("500", "500", "500", "500", "500",
+                "300", "300", "280", "280", "280"));
+    }
+
+    @Test(dependsOnMethods = {"testRestoreFromFileSystem"})
+    public void testFileSystemPeriodicPersistence() throws InterruptedException {
+        File file = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
+        log.info("Waiting for third time interval for state persistence");
+        Thread.sleep(60000); //await() cannot be used because number of persisted revisions do not change
+
+        Assert.assertEquals(file.exists() && file.isDirectory(), true);
+        Assert.assertEquals(file.list().length, 2, "There should be two revisions persisted");
+    }
+
+    @Test(dependsOnMethods = {"testFileSystemPeriodicPersistence"})
+    public void testFileSystemClearPeriodicPersistence() {
+        File file = new File(PERSISTENCE_FOLDER + File.separator + SIDDHIAPP_NAME);
+
+        SiddhiManager siddhiManager = StreamProcessorDataHolder.getSiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.getSiddhiAppRuntime(SIDDHIAPP_NAME);
+
+        if (file.list().length == 0) {
+            Assert.fail("File revisions are already deleted");
+        }
+
+        log.info("Deleting all the revisions of the persistence store of Siddhi App : " + SIDDHIAPP_NAME );
+        siddhiAppRuntime.clearAllRevisions();
+        Assert.assertEquals(file.list().length, 0, "All the revisions should be deleted");
+    }
+
 }
