+++ osgi-tests/src/test/java/io/siddhi/distribution/test/osgi/util/TestUtil.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.wso2.carbon.analytics.test.osgi.util;
+package io.siddhi.distribution.test.osgi.util;
 
 import io.netty.handler.codec.http.HttpMethod;
+import io.siddhi.distribution.common.common.EventStreamService;
+import io.siddhi.distribution.common.common.SiddhiAppRuntimeService;
 import org.awaitility.Duration;
-import org.wso2.carbon.stream.processor.common.EventStreamService;
-import org.wso2.carbon.stream.processor.common.SiddhiAppRuntimeService;
 import org.wso2.msf4j.MicroservicesRegistry;
 import org.wso2.siddhi.core.SiddhiAppRuntime;
