+++ osgi-tests/src/test/resources/conf/persistence/db/deployment-structure.yaml
@@ -256,7 +256,7 @@ state.persistence:
   enabled: true
   intervalInMin: 1
   revisionsToKeep: 2
-  persistenceStore: org.wso2.carbon.stream.processor.core.persistence.DBPersistenceStore
+  persistenceStore: io.siddhi.distribution.core.persistence.DBPersistenceStore
   config:
     datasource: WSO2_ANALYTICS_DB
     table: PERSISTENCE_TABLE
