+++ osgi-tests/src/test/resources/conf/persistence/file/configTest/deployment.yaml
@@ -241,7 +241,7 @@ wso2.artifact.deployment:
   # HA Configuration
 state.persistence:
   enabled: true
-  persistenceStore: org.wso2.carbon.stream.processor.core.persistence.FileSystemPersistenceStore
+  persistenceStore: io.siddhi.distribution.core.persistence.FileSystemPersistenceStore
   config:
 
   # Secure Vault Configuration
