+++ osgi-tests/src/test/resources/conf/persistence/file/default/deployment.yaml
@@ -273,7 +273,7 @@ state.persistence:
   enabled: true
   intervalInMin: 1
   revisionsToKeep: 2
-  persistenceStore: org.wso2.carbon.stream.processor.core.persistence.FileSystemPersistenceStore
+  persistenceStore: io.siddhi.distribution.core.persistence.FileSystemPersistenceStore
   config:
     location: siddhi-app-persistence
