+++ osgi-tests/src/test/resources/conf/persistence/incremental-db/deployment-structure.yaml
@@ -230,7 +230,7 @@ state.persistence:
   enabled: true
   intervalInMin: 1
   revisionsToKeep: 2
-  persistenceStore: org.wso2.carbon.stream.processor.core.persistence.IncrementalDBPersistenceStore
+  persistenceStore: io.siddhi.distribution.core.persistence.IncrementalDBPersistenceStore
   config:
     datasource: WSO2_ANALYTICS_DB
     table: PERSISTENCE_TABLE
