+++ osgi-tests/src/test/resources/testng.xml
@@ -14,25 +14,20 @@ limitations under the License.
 
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="Carbon-Analytics-Test-Suite">
-    <test name="carbon-analytics-tests" preserve-order="true" parallel="false" thread-count="1">
+<suite name="Siddhi-Distribution-Test-Suite">
+    <test name="siddhi-distribution-tests" preserve-order="true" parallel="false" thread-count="1">
         <classes>
-            <class name="org.wso2.carbon.analytics.test.osgi.SiddhiMetricsAPITestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.FileSystemPersistenceStoreTestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.IncrementalFileSystemPersistenceStoreTestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.SiddhiAsAPITestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.SiddhiMetricsTestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.SiddhiStoreAPITestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.DBPersistenceStoreTestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.IncrementalDBPersistenceStoreTestcase"/>
-            <!--<class name="org.wso2.carbon.analytics.test.osgi.HAActiveNodeTestcase"/>-->
-            <class name="org.wso2.carbon.analytics.test.osgi.FileSystemPersistenceStoreConfigTestcase"/>
-            <class name="org.wso2.carbon.analytics.test.osgi.SiddhiEditorTestCase"/>
-            <!--<class name="org.wso2.carbon.analytics.test.osgi.DistributedJobManagerTestcase"/>-->
-            <!--<class name="org.wso2.carbon.analytics.test.osgi.DistributedResourceTestcase"/>-->
-            <class name="org.wso2.carbon.analytics.test.osgi.SimulatorAPITestcase"/>
-            <!--<class name="org.wso2.carbon.analytics.test.osgi.StatusDashboardWorkerTestCase"/>-->
-            <class name="org.wso2.carbon.analytics.test.osgi.ConfigReaderTestCase"/>
+            <class name="io.siddhi.distribution.test.osgi.FileSystemPersistenceStoreTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.IncrementalFileSystemPersistenceStoreTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.SiddhiAsAPITestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.SiddhiMetricsTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.SiddhiStoreAPITestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.DBPersistenceStoreTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.IncrementalDBPersistenceStoreTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.FileSystemPersistenceStoreConfigTestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.SiddhiEditorTestCase"/>
+            <class name="io.siddhi.distribution.test.osgi.SimulatorAPITestcase"/>
+            <class name="io.siddhi.distribution.test.osgi.ConfigReaderTestCase"/>
         </classes>
     </test>
 </suite>
