+++ osgi-tests/src/test/resources/testngdocker.xml
@@ -17,7 +17,7 @@ limitations under the License.
 <suite name="Carbon-Analytics-Test-Suite">
     <test name="carbon-analytics-tests" preserve-order="true">
         <classes>
-            <class name="org.wso2.carbon.analytics.test.osgi.DBPersistenceStoreTestcase"/>
+            <class name="DBPersistenceStoreTestcase"/>
         </classes>
     </test>
 </suite>
