+++ runner/components/io.siddhi.distribution.event.simulator.core/pom.xml
@@ -22,14 +22,14 @@
         <groupId>io.siddhi.distribution</groupId>
         <artifactId>io.siddhi.distribution.parent</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
     <artifactId>io.siddhi.distribution.event.simulator.core</artifactId>
     <packaging>bundle</packaging>
 
-    <name>WSO2 Stream Processor - Event Simulator</name>
+    <name>Siddhi Distribution - Event Simulator</name>
     <url>http://wso2.org</url>
 
     <dependencies>
