+++ runner/components/io.siddhi.distribution.metrics.core/pom.xml
@@ -22,12 +22,12 @@
         <groupId>io.siddhi.distribution</groupId>
         <artifactId>io.siddhi.distribution.parent</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
     <artifactId>io.siddhi.distribution.metrics.core</artifactId>
-    <name>WSO2 Carbon - Siddhi Metrics Core</name>
+    <name>Siddhi Distribution - Metrics Core</name>
     <packaging>bundle</packaging>
 
     <dependencies>
