+++ runner/components/io.siddhi.distribution.msf4j.interceptor.common/pom.xml
@@ -22,13 +22,13 @@
         <groupId>io.siddhi.distribution</groupId>
         <artifactId>io.siddhi.distribution.parent</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
     </parent>
 
     <artifactId>io.siddhi.distribution.msf4j.interceptor.common</artifactId>
     <packaging>bundle</packaging>
 
-    <name>WSO2 Stream Processor - Interceptors</name>
+    <name>Siddhi Distribution - Interceptors</name>
     <url>http://maven.apache.org</url>
 
     <dependencies>
