+++ runner/components/io.siddhi.distribution.store.api.rest/pom.xml
@@ -23,14 +23,14 @@
         <groupId>io.siddhi.distribution</groupId>
         <artifactId>io.siddhi.distribution.parent</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
     <artifactId>io.siddhi.distribution.store.api.rest</artifactId>
     <packaging>bundle</packaging>
 
-    <name>WSO2 Stream Processor - Siddhi Store Query REST API</name>
+    <name>Siddhi Distribution - Store Query REST API</name>
     <url>http://wso2.org</url>
 
     <dependencies>
